+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1062](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1062)
 - Implemented counts endpoint for sensors, streams, datapoints and bins
   [GEOD-1154](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1154)
+- Added logs for start and end time in binning endpoints
+  [GEOD-1162](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1162)
   
 ### Changed
 - Changed header and email title